feat: include assertionUtils in test-utils package
And use for redis and pg
