--- packages/opentelemetry-plugin-postgres/opentelemetry-plugin-pg-pool/test/assertionUtils.ts
@@ -1,79 +1,0 @@
-/*!
- * Copyright 2019, OpenTelemetry Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import {
-  SpanKind,
-  Attributes,
-  Event,
-  Span,
-  TimedEvent,
-} from '@opentelemetry/types';
-import * as assert from 'assert';
-import { ReadableSpan } from '@opentelemetry/tracing';
-import {
-  hrTimeToMilliseconds,
-  hrTimeToMicroseconds,
-} from '@opentelemetry/core';
-
-export const assertSpan = (
-  span: ReadableSpan,
-  kind: SpanKind,
-  attributes: Attributes,
-  events: Event[]
-) => {
-  assert.strictEqual(span.spanContext.traceId.length, 32);
-  assert.strictEqual(span.spanContext.spanId.length, 16);
-  assert.strictEqual(span.kind, kind);
-
-  // check all the AttributeNames fields
-  Object.keys(span.attributes).forEach(key => {
-    assert.deepStrictEqual(span.attributes[key], attributes[key]);
-  });
-
-  assert.ok(span.endTime);
-  assert.strictEqual(span.links.length, 0);
-
-  assert.ok(
-    hrTimeToMicroseconds(span.startTime) < hrTimeToMicroseconds(span.endTime)
-  );
-  assert.ok(hrTimeToMilliseconds(span.endTime) > 0);
-
-  // events
-  assert.strictEqual(
-    span.events.length,
-    events.length,
-    'Should contain same number of events'
-  );
-  span.events.forEach((_: TimedEvent, index: number) => {
-    assert.deepStrictEqual(span.events[index], events[index]);
-  });
-};
-
-// Check if sourceSpan was propagated to targetSpan
-export const assertPropagation = (
-  childSpan: ReadableSpan,
-  parentSpan: Span
-) => {
-  const targetSpanContext = childSpan.spanContext;
-  const sourceSpanContext = parentSpan.context();
-  assert.strictEqual(targetSpanContext.traceId, sourceSpanContext.traceId);
-  assert.strictEqual(childSpan.parentSpanId, sourceSpanContext.spanId);
-  assert.strictEqual(
-    targetSpanContext.traceFlags,
-    sourceSpanContext.traceFlags
-  );
-  assert.notStrictEqual(targetSpanContext.spanId, sourceSpanContext.spanId);
-};
