+++ packages/opentelemetry-test-utils/package.json
@@ -27,5 +27,8 @@
     "tslint-consistent-codestyle": "^1.16.0",
     "tslint-microsoft-contrib": "^6.2.0",
     "typescript": "3.7.4"
+  },
+  "dependencies": {
+    "@opentelemetry/tracing": "^0.3.1"
   }
 }
