+++ packages/opentelemetry-test-utils/testUtils.ts
@@ -15,6 +15,20 @@
  */
 
 import * as childProcess from 'child_process';
+import {
+  SpanKind,
+  Attributes,
+  Event,
+  Span,
+  Status,
+} from '@opentelemetry/types';
+import * as assert from 'assert';
+import { ReadableSpan } from '@opentelemetry/tracing';
+import {
+  hrTimeToMilliseconds,
+  hrTimeToMicroseconds,
+} from '@opentelemetry/core';
+
 export function startDocker(db: 'redis' | 'mysql' | 'postgres') {
   let dockerRunCmd;
   switch (db) {
@@ -65,3 +79,50 @@ function run(cmd: string) {
     return;
   }
 }
+
+export const assertSpan = (
+  span: ReadableSpan,
+  kind: SpanKind,
+  attributes: Attributes,
+  events: Event[],
+  status: Status
+) => {
+  assert.strictEqual(span.spanContext.traceId.length, 32);
+  assert.strictEqual(span.spanContext.spanId.length, 16);
+  assert.strictEqual(span.kind, kind);
+
+  assert.ok(span.endTime);
+  assert.strictEqual(span.links.length, 0);
+
+  assert.ok(
+    hrTimeToMicroseconds(span.startTime) < hrTimeToMicroseconds(span.endTime)
+  );
+  assert.ok(hrTimeToMilliseconds(span.endTime) > 0);
+
+  // attributes
+  assert.deepStrictEqual(span.attributes, attributes);
+
+  // events
+  assert.deepStrictEqual(span.events, events);
+
+  assert.strictEqual(span.status.code, status.code);
+  if (status.message) {
+    assert.strictEqual(span.status.message, status.message);
+  }
+};
+
+// Check if childSpan was propagated from parentSpan
+export const assertPropagation = (
+  childSpan: ReadableSpan,
+  parentSpan: Span
+) => {
+  const targetSpanContext = childSpan.spanContext;
+  const sourceSpanContext = parentSpan.context();
+  assert.strictEqual(targetSpanContext.traceId, sourceSpanContext.traceId);
+  assert.strictEqual(childSpan.parentSpanId, sourceSpanContext.spanId);
+  assert.strictEqual(
+    targetSpanContext.traceFlags,
+    sourceSpanContext.traceFlags
+  );
+  assert.notStrictEqual(targetSpanContext.spanId, sourceSpanContext.spanId);
+};
